✨ Add games from 2020 and before
Found these in an old folder - they were on the original "blue" design of natelev.in, but never made it past the "red" redesign in 2020. Now they're back!

You'll notice that the code is quite low-quality - I used these projects to learn JavaScript (and programming in general!) back in 2019 and early 2020.